Extract log record search into a separate module too
Issue: #297
