+++ search/logrecord/src/main/java/name/mlopatkin/andlogview/search/logrecord/RowSearchStrategy.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 Mikhail Lopatkin
+ * Copyright 2013 Mikhail Lopatkin
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,12 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package name.mlopatkin.andlogview.search.logrecord;
 
-rootProject.name = "andlogview"
+import name.mlopatkin.andlogview.logmodel.Field;
+import name.mlopatkin.andlogview.logmodel.LogRecord;
+import name.mlopatkin.andlogview.search.text.TextHighlighter;
 
-include("base")
-include("device")
-include("logmodel")
-include("parsers")
-include("search")
-include("search:logrecord")
+import java.util.function.Predicate;
+
+public interface RowSearchStrategy extends Predicate<LogRecord> {
+    void highlightColumn(LogRecord record, Field field, TextHighlighter columnHighlighter);
+}
