+++ search/logrecord/src/test/java/name/mlopatkin/andlogview/search/logrecord/RowSearchersTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package name.mlopatkin.andlogview.search;
+package name.mlopatkin.andlogview.search.logrecord;
 
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
