+++ src/name/mlopatkin/andlogview/ui/search/logtable/SearchResultsHighlightCellRenderer.java
@@ -17,7 +17,7 @@ package name.mlopatkin.andlogview.ui.search.logtable;
 
 import name.mlopatkin.andlogview.TooltipGenerator;
 import name.mlopatkin.andlogview.logmodel.LogRecord;
-import name.mlopatkin.andlogview.search.RowSearchStrategy;
+import name.mlopatkin.andlogview.search.logrecord.RowSearchStrategy;
 import name.mlopatkin.andlogview.search.text.TextHighlighter;
 import name.mlopatkin.andlogview.ui.logtable.Column;
 import name.mlopatkin.andlogview.ui.logtable.LogRecordTableModel;
